+++ src/Template/TemplateRendererInterface.php
@@ -15,6 +15,11 @@ namespace Zend\Expressive\Template;
 interface TemplateRendererInterface
 {
     /**
+     * @const string Value indicating all templates; used with `addDefaultParam()`.
+     */
+    const TEMPLATE_ALL = '*';
+
+    /**
      * Render a template, optionally with parameters.
      *
      * Implementations MUST support the `namespace::template` naming convention,
@@ -43,4 +48,25 @@ interface TemplateRendererInterface
      * @return TemplatePath[]
      */
     public function getPaths();
+
+    /**
+     * Add a default parameter to use with a template.
+     *
+     * Use this method to provide a default parameter to use when a template is
+     * rendered. The parameter may be overridden by providing it when calling
+     * `render()`, or by calling this method again with a null value.
+     *
+     * The parameter will be specific to the template name provided. To make
+     * the parameter available to any template, pass the TEMPLATE_ALL constant
+     * for the template name.
+     *
+     * If the default parameter existed previously, subsequent invocations with
+     * the same template name and parameter name will overwrite.
+     *
+     * @param string $templateName Name of template to which the param applies;
+     *     use TEMPLATE_ALL to apply to all templates.
+     * @param string $param Param name.
+     * @param mixed $value
+     */
+    public function addDefaultParam($templateName, $param, $value);
 }