+++ test/Template/TestAsset/plates-2.php
@@ -1,0 +1,3 @@
+<h1>This is a second template file for Plates</h1>
+
+<p>You are using <strong><?=$this->e($name)?></strong>!</p>